fix type field

+++ database/migrations/2021_11_21_173516_create_authors_table.php
@@ -18,7 +18,7 @@ class CreateAuthorsTable extends Migration
             $table->string('name')->nullable();
             $table->string('surname')->nullable();
             $table->string('patronymic')->nullable();
-            $table->integer('age')->nullable();
+            $table->unsignedBigInteger('age')->nullable();
             $table->timestamps();
         });
     }
